chore: fix gosec G115 lint

+++ internal/pkg/util/user/identity_unix.go
@@ -59,5 +59,5 @@ func CurrentOriginal() (*User, error) {
 	if err != nil {
 		return nil, err
 	}
-	return GetPwUID(uint32(uid))
+	return GetPwUID(uid)
 }
